bot/requirements.txt
python-telegram-bot==20.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.6.1
pydantic==2.5.3
pydantic-settings==2.1.0
